Refactor:  useVMTimer 로직 VMTimerProvider로 합치기
issue #13
