+++ src/Context/VMTimerProvider.jsx
@@ -1,12 +1,30 @@
 /* eslint-disable react/prop-types */
-import React, { createContext, useMemo } from 'react';
-import useVMTimer from 'hooks/useVMTimer';
+import React, { useState, createContext, useCallback, useMemo } from 'react';
 
 export const VMTimerContext = createContext([]);
 export const VMTimerSetContext = createContext(null);
 
 export function VMTimerProvider({ children }) {
-  const { VMTimerID, startVMTimer, stopVMTimer } = useVMTimer();
+  const [VMTimerID, setVMTimerId] = useState([]);
+
+  const startVMTimer = useCallback(callbackArr => {
+    const [newVMTimerID] = callbackArr.reduce(
+      ([timerIDArr, accTime], [callback, time]) => {
+        const timerID = setTimeout(callback, accTime + time);
+        return [[...timerIDArr, timerID], accTime + time];
+      },
+      [[], 0]
+    );
+    setVMTimerId(newVMTimerID);
+  }, []);
+
+  const stopVMTimer = useCallback(() => {
+    if (!VMTimerID) return;
+
+    VMTimerID.forEach(ID => clearTimeout(ID));
+    setVMTimerId([]);
+  }, [VMTimerID]);
+
   const setVMTimerID = useMemo(() => ({ startVMTimer, stopVMTimer }), [startVMTimer, stopVMTimer]);
 
   return (
